Added comment
Clarification on how to tailor. I don't know batch scripting yet so I couldn't make it dynamic or even dependent on one variable. However for the time being it is simple enough to do.

This is, with the assumption that xlaunch files can work on different systems as long as VcXsrv is installed.